schedule_execution_monitor: Includes scheduler
